completed controllers and views

+++ db/seeds.rb
@@ -1,20 +1,15 @@
-# This file should contain all the record creation needed to seed the database with its default values.
-# The data can then be loaded with the rails db:seed command (or created alongside the database with db:setup).
-#
-# Examples:
-#
-#   movies = Movie.create([{ name: 'Star Wars' }, { name: 'Lord of the Rings' }])
-#   Character.create(name: 'Luke', movie: movies.first)
-require 'json'
 require 'open-uri'
-Ingredient.destroy_all
 
-puts "Creating ingredients ..."
+puts "Destroy ingredients"
+Ingredient.destroy_all if Rails.env.development?
 
-url = 'https://www.thecocktaildb.com/api/json/v1/1/list.php?i=list'
+puts "Destroy Cocktails"
+Cocktail.destroy_all if Rails.env.development?
+
+puts "Create ingredients"
+url = "https://www.thecocktaildb.com/api/json/v1/1/list.php?i=list"
 ingredients = JSON.parse(open(url).read)
-
 ingredients["drinks"].each do |ingredient|
-  Ingredient.create(name: ingredient["strIngredient"])
+  i = Ingredient.create(name: ingredient["strIngredient1"])
+  puts "create #{i.name}"
 end
-puts "Finished"
